Refactor metadata to blueprint/service
Fixes #1992

This change migrates the existing /<country>/metadata route to a
blueprint. This involved
1. Creating a blueprint for just metadata which adds the
   /<country>/metadata route
2. Updating the get_metadata function to use a MetadataService instead
   of directly referencing COUNTRIES.

Existing build tests already cover this functionality, but I also ran
debug and verified with curl that the route responded appropriately.
