+++ policyengine_api/routes/metadata_routes.py
@@ -1,0 +1,18 @@
+from flask import Blueprint
+
+from policyengine_api.utils.payload_validators import validate_country
+from policyengine_api.services.metadata_service import MetadataService
+
+metadata_bp = Blueprint("metadata", __name__)
+metadata_service = MetadataService()
+
+
+@metadata_bp.route("/<country_id>/metadata", methods=["GET"])
+@validate_country
+def get_metadata(country_id: str) -> dict:
+    """Get metadata for a country.
+
+    Args:
+        country_id (str): The country ID.
+    """
+    return metadata_service.get_metadata(country_id)
